refactor PatternPage folder for cleaner encapsulation
