+++ polaris.shopify.com/src/components/PatternPage/index.ts
@@ -1,0 +1,2 @@
+import PatternPage from './PatternPage';
+export default PatternPage;
