--- polaris.shopify.com/src/components/PatternPage/index.tsx
@@ -1,4 +1,0 @@
-import SingleVariantPatternPage from './SingleVariantPatternPage';
-import MultiVariantPatternPage from './MultiVariantPatternPage';
-
-export {SingleVariantPatternPage, MultiVariantPatternPage};
